File Upload & Download 1. Image file upload and its storage in local drive 2. Invoice generation of orders.

+++ controllers/admin.js
@@ -1,6 +1,10 @@
-const {validationResult}=require('express-validator/check');
+const { validationResult } = require('express-validator/check');
 const Product = require('../models/product');
 const { error } = require('console');
+
+const fileHelper = require('../util/file');
+const product = require('../models/product');
+
 
 
 exports.getAddProduct = (req, res, next) => {
@@ -9,32 +13,60 @@ exports.getAddProduct = (req, res, next) => {
     pageTitle: 'Add Product',
     path: '/admin/add-product',
     editing: false,
-    hasError:false,
-    errorMessage:null,
+    hasError: false,
+    errorMessage: null,
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
+  // console.log('adding Product');
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
 
-  const errors=validationResult(req);
+  // console.log(image)
 
-  if(!errors.isEmpty()){
+  if (!image) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/edit-product',
       editing: false,
-      hasError:true,
+      hasError: true,
       product: {
-        title:title,
-        imageUrl:imageUrl,
-        price:price,
-        description:description,
+        title: title,
+        price: price,
+        description: description,
       },
-      errorMessage:errors.array()[0].msg,
+      errorMessage: "Attached file is not an image.",
+
+
+    });
+  }
+
+
+  const imageUrl = image.path;
+
+  // console.log(imageUrl);
+
+  // console.log(req);
+
+
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        imageUrl: imageUrl,
+        price: price,
+        description: description,
+      },
+      errorMessage: errors.array()[0].msg,
 
     });
   }
@@ -55,8 +87,9 @@ exports.postAddProduct = (req, res, next) => {
     })
     .catch(err => {
       // console.log(err);
-      const error=new Error(err);
-      error.httpStatusCode(500);
+      const error = new Error(err);
+
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -77,14 +110,14 @@ exports.getEditProduct = (req, res, next) => {
         path: '/admin/edit-product',
         editing: editMode,
         product: product,
-        hasError:false,
-        errorMessage:null,
+        hasError: false,
+        errorMessage: null,
       });
     })
-    .catch(err =>{
+    .catch(err => {
       // console.log(err);
-      const error=new Error(err);
-      error.httpStatusCode(500);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -93,32 +126,31 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const updatedImage = req.file;
   const updatedDesc = req.body.description;
 
-  const errors=validationResult(req);
+  const errors = validationResult(req);
 
-  if(!errors.isEmpty()){
+  if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Edit Product',
       path: '/admin/edit-product',
       editing: true,
-      hasError:true,
+      hasError: true,
       product: {
-        _id:prodId,
-        title:updatedTitle,
-        imageUrl:updatedImageUrl,
-        price:updatedPrice,
-        description:updatedDesc,
+        _id: prodId,
+        title: updatedTitle,
+        price: updatedPrice,
+        description: updatedDesc,
       },
-      errorMessage:errors.array()[0].msg, 
+      errorMessage: errors.array()[0].msg,
 
     });
   }
 
   Product.findById(prodId)
     .then(product => {
-      
+
       console.log(req.user._id);
       if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect('/');
@@ -127,7 +159,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (updatedImage) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = updatedImage.path;
+      }
       return product.save()
         .then(result => {
           console.log('UPDATED PRODUCT!');
@@ -137,8 +172,8 @@ exports.postEditProduct = (req, res, next) => {
 
     .catch(err => {
       // console.log(err);
-      const error=new Error(err);
-      error.httpStatusCode(500);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -156,23 +191,30 @@ exports.getProducts = (req, res, next) => {
     })
     .catch(err => {
       // console.log(err);
-      const error=new Error(err);
-      error.httpStatusCode(500);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId:req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
     .catch(err => {
       // console.log(err);
-      const error=new Error(err);
-      error.httpStatusCode(500);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
